GCP-handbook/gcp-services/compute/requirements.txt
google-cloud-compute>=1.14.1
google-cloud-container>=2.20.0
google-cloud-run>=0.10.0
google-cloud-appengine>=1.12.0
google-cloud-functions>=1.13.0
kubernetes>=28.1.0
flask>=2.3.0
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
